+++ package.json
@@ -13,6 +13,7 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.19.0",
+    "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.17.1",
     "mssql": "10.0.3",
